requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
environs==14.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
